# frontend/mrsexam-ui/root/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "mrsexam-backend",
    "version": "1.0"
  },
  "paths": {
    "/api/AttSheets": {
      "get": {
        "tags": [
          "AttSheets"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "AttSheets"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttSheet"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AttSheet"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AttSheet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/AttSheets/{id}": {
      "get": {
        "tags": [
          "AttSheets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "AttSheets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttSheet"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AttSheet"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AttSheet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "AttSheets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/AttSheets/centre/{examCentre}": {
      "get": {
        "tags": [
          "AttSheets"
        ],
        "parameters": [
          {
            "name": "examCentre",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Centers": {
      "get": {
        "tags": [
          "Centers"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Centers"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Center"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Center"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Center"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Centers/paged": {
      "get": {
        "tags": [
          "Centers"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Centers/{id}": {
      "get": {
        "tags": [
          "Centers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Centers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Center"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Center"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Center"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Centers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CentreCounts": {
      "get": {
        "tags": [
          "CentreCounts"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "CentreCounts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CentreCount"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CentreCount"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CentreCount"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CentreCounts/{id}": {
      "get": {
        "tags": [
          "CentreCounts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "CentreCounts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CentreCount"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CentreCount"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CentreCount"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "CentreCounts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CentreCounts/centre/{centreNo}": {
      "get": {
        "tags": [
          "CentreCounts"
        ],
        "parameters": [
          {
            "name": "centreNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CentreCounts/region/{regionNo}": {
      "get": {
        "tags": [
          "CentreCounts"
        ],
        "parameters": [
          {
            "name": "regionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormBalbodhini/centre/{centreCode}": {
      "get": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "centreCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormPandit/centre/{centreCode}": {
      "get": {
        "tags": [
          "FormPandit"
        ],
        "parameters": [
          {
            "name": "centreCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormPrabodh/centre/{centreCode}": {
      "get": {
        "tags": [
          "FormPrabodh"
        ],
        "parameters": [
          {
            "name": "centreCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormPrathamik/centre/{centreCode}": {
      "get": {
        "tags": [
          "FormPrathamik"
        ],
        "parameters": [
          {
            "name": "centreCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormPraveen/centre/{centreCode}": {
      "get": {
        "tags": [
          "FormPraveen"
        ],
        "parameters": [
          {
            "name": "centreCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormPraveshika/centre/{centreCode}": {
      "get": {
        "tags": [
          "FormPraveshika"
        ],
        "parameters": [
          {
            "name": "centreCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/FormSubodh/centre/{centreCode}": {
      "get": {
        "tags": [
          "FormSubodh"
        ],
        "parameters": [
          {
            "name": "centreCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttSheet": {
        "type": "object",
        "properties": {
          "attSheetId": {
            "type": "integer",
            "format": "int32"
          },
          "examCentre": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "col1RollNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "col2RollNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Center": {
        "required": [
          "centreNo"
        ],
        "type": "object",
        "properties": {
          "centreNo": {
            "maxLength": 10,
            "minLength": 1,
            "type": "string"
          },
          "addressTo": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "centreName": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "districtName": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "add1": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "add2": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "city": {
            "maxLength": 35,
            "type": "string",
            "nullable": true
          },
          "pincode": {
            "maxLength": 6,
            "type": "string",
            "nullable": true
          },
          "phone": {
            "maxLength": 15,
            "type": "string",
            "nullable": true
          },
          "panditFlag": {
            "maxLength": 1,
            "type": "string",
            "nullable": true
          },
          "closeFlag": {
            "maxLength": 1,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CentreCount": {
        "type": "object",
        "properties": {
          "centreCountId": {
            "type": "integer",
            "format": "int32"
          },
          "centreNo": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "centreName": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "fromGroup1": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "toGroup1": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "fromGroup2": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "toGroup2": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "other": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "regionNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}